python/requirements.txt
ollama
pyyaml
rich
sqlalchemy
sentence-transformers
numpy
